Further changes to GsonSubTypeExtension
