+++ compiler/standard/src/main/java/gsonpath/generator/extension/subtype/GsonSubTypeResult.kt
@@ -1,0 +1,14 @@
+package gsonpath.generator.extension.subtype
+
+import com.squareup.javapoet.CodeBlock
+import com.squareup.javapoet.FieldSpec
+import com.squareup.javapoet.MethodSpec
+import com.squareup.javapoet.TypeSpec
+
+data class GsonSubTypeResult(
+        val constructorCodeBlock: CodeBlock? = null,
+        val fieldSpecs: List<FieldSpec> = emptyList(),
+        val typeSpecs: List<TypeSpec> = emptyList(),
+        val readMethodSpecs: MethodSpec,
+        val writeMethodSpecs: MethodSpec
+)
